mfcdec: Compensate race condition in video thumbnail generation
Calling the decoder immediately after encoding a video often
trips on the ongoing mfc buffer flushes from the encoder that has
just shut down, and fails due to the unavailability of any buffer of
sufficient size. If MFC_DEC_INIT fails, give it 100msec and try
again. This resolves the blank thumbs when switching to the
gallery in the same session that a recording was made.

Change-Id: I81b78c807b28cccef7a99f1106989754ea39fb97